+++ lol/ViGEmXbox360Output.cs
@@ -31,8 +31,8 @@ main
         {
             try
             {
-                _controller.SetAxisValue(Xbox360Axis.RightThumbX, x);
-                _controller.SetAxisValue(Xbox360Axis.RightThumbY, y);
+                _controller.SetAxisValue(Xbox360Axis.RightThumbX, (short)x);
+                _controller.SetAxisValue(Xbox360Axis.RightThumbY, (short)y);
 > main
                 _controller.SubmitReport();
             }
